[jak3] Set up ckernel (#3308)
This sets up the C Kernel for Jak 3, and makes it possible to build and
load code built with `goalc --jak3`.

There's not too much interesting here, other than they switched to a
system where symbol IDs (unique numbers less than 2^14) are generated at
compile time, and those get included in the object file itself.

This is kind of annoying, since it means all tools that produce a GOAL
object file need to work together to assign unique symbol IDs. And since
the symbol IDs can't conflict, and are only a number between 0 and 2^14,
you can't just hash and hope for no collisions.

We work around this by ignoring the IDs and re-assigning our own. I
think this is very similar to what the C Kernel did on early builds of
Jak 3 which supported loading old format level files, which didn't have
the IDs included.

As far as I can tell, this shouldn't cause any problems. It defeats all
of their fancy tricks to save memory by not storing the symbol string,
but we don't care.

+++ decompiler/ObjectFile/ObjectFileDB_IR2.cpp
@@ -367,7 +367,7 @@ void ObjectFileDB::ir2_analyze_all_types(const fs::path& output_file,
         for (const auto& [obj_name, type_names] : all_type_names) {
           for (const auto& type_name : type_names) {
             if (str_util::starts_with(sym_name, type_name) &&
-                type_name.length() > longest_match_length) {
+                (int)type_name.length() > longest_match_length) {
               longest_match_length = type_name.length();
               longest_match = type_name;
               longest_match_object_name = obj_name;
